Move software updater into it's own module
